Enum easy_plugin::Match

pub enum Match {
    Attr(Attribute),
    BinOp(BinOpToken),
    Block(P<Block>),
    Delim(Rc<Delimited>),
    Expr(P<Expr>),
    Ident(Ident),
    Item(P<Item>),
    Lftm(Name),
    Lit(Lit),
    Meta(P<MetaItem>),
    Pat(P<Pat>),
    Path(Path),
    Stmt(Stmt),
    Ty(P<Ty>),
    Tok(Token),
    Tt(TokenTree),
    Sequence(Vec<Match>),
    NamedSequence(usize),
}

A plugin argument that has been matched with a named specifier.

Variants

Attr

An attribute (e.g., #[cfg(target_os = "windows")]).

BinOp

A binary operator (e.g., +, *).

Block

A brace-delimited sequence of statements (e.g., { log(error, "hi"); return 12; }).

Delim

A delimited sequence of token trees (e.g., (), [foo - "bar"]).

Expr

An expression (e.g., 2 + 2, if true { 1 } else { 2 }, f(42)).

Ident

An identifier (e.g., x, foo).

Item

An item (e.g., fn foo() { }, struct Bar;).

Lftm

A lifetime (e.g., 'a).

Lit

A literal (e.g., 322, 'a', "foo").

Meta

A "meta" item, as found in attributes (e.g., cfg(target_os = "windows")).

Pat

A pattern (e.g., Some(t), (17, 'a'), _).

Path

A qualified name (e.g., T::SpecialA).

Stmt

A single statement (e.g., let x = 3).

Ty

A type (e.g., i32, Vec<(char, String)>, &T).

Tok

A single token.

Tt

A single token tree.

Sequence

A sequence which may either contain sequence matches or subsequences.

NamedSequence

A count of named sequence repetitions.

Methods

impl Match

fn as_attr(&self) -> Attribute

Returns this attribute match.

Panics

this match is not an attribute

fn as_binop(&self) -> BinOpToken

Returns this binary operator match.

Panics

this match is not a binary operator

fn as_block(&self) -> P<Block>

Returns this block match.

Panics

this match is not a block

fn as_delim(&self) -> Rc<Delimited>

Returns this delimited match.

Panics

this match is not a delimited sequence of token trees

fn as_expr(&self) -> P<Expr>

Returns this expression match.

Panics

this match is not an expression

fn as_ident(&self) -> Ident

Returns this identifier match.

Panics

this match is not an identifier

fn as_item(&self) -> P<Item>

Returns this item match.

Panics

this match is not an item

fn as_lftm(&self) -> Name

Returns this lifetime match.

Panics

this match is not a lifetime

fn as_lit(&self) -> Lit

Returns this literal match.

Panics

this match is not a literal

fn as_meta(&self) -> P<MetaItem>

Returns this "meta" item match.

Panics

this match is not a "meta" item

fn as_pat(&self) -> P<Pat>

Returns this pattern match.

Panics

this match is not a pattern

fn as_path(&self) -> Path

Returns this path match.

Panics

this match is not a path

fn as_stmt(&self) -> Stmt

Returns this statement match.

Panics

this match is not a statement

fn as_ty(&self) -> P<Ty>

Returns this type match.

Panics

this match is not a type

fn as_tok(&self) -> Token

Returns this token match.

Panics

this match is not a token

fn as_tt(&self) -> TokenTree

Returns this token tree match.

Panics

this match is not a token tree

fn as_sequence(&self) -> Vec<Match>

Returns this collection of sequence matches or subsequences.

Panics

this match is not a collection of sequence matches or subsequences

fn as_named_sequence(&self) -> usize

Returns this count of named sequence repetitions.

Panics

this match is not a count of named sequence reptitions

fn as_named_sequence_bool(&self) -> bool

Returns whether this count of named sequence repetitions is non-zero.

Panics

this match is not a count of named sequence reptitions