Enum easy_plugin::Specifier

pub enum Specifier {
    Attr(String),
    BinOp(String),
    Block(String),
    Delim(String),
    Expr(String),
    Ident(String),
    Item(String),
    Lftm(String),
    Lit(String),
    Meta(String),
    Pat(String),
    Path(String),
    Stmt(String),
    Ty(String),
    Tok(String),
    Tt(String),
    Specific(Token),
    Delimited(DelimToken, Specification),
    Sequence(Amount, Option<Token>, Specification),
    NamedSequence(String, Amount, Option<Token>, Specification),
}

A piece of a plugin argument specification.

Variants

Attr

An attribute (e.g., #[cfg(target_os = "windows")]).

BinOp

A binary operator (e.g., +, *).

Block

A brace-delimited sequence of statements (e.g., { log(error, "hi"); return 12; }).

Delim

A delimited sequence of token trees (e.g., (), [foo - "bar"]).

Expr

An expression (e.g., 2 + 2, if true { 1 } else { 2 }, f(42)).

Ident

An identifier (e.g., x, foo).

Item

An item (e.g., fn foo() { }, struct Bar;).

Lftm

A lifetime (e.g., 'a).

Lit

A literal (e.g., 322, 'a', "foo").

Meta

A "meta" item, as found in attributes (e.g., cfg(target_os = "windows")).

Pat

A pattern (e.g., Some(t), (17, 'a'), _).

Path

A qualified name (e.g., T::SpecialA).

Stmt

A single statement (e.g., let x = 3).

Ty

A type (e.g., i32, Vec<(char, String)>, &T).

Tok

A single token.

Tt

A single token tree.

Specific

A non-variable piece.

Delimited

A delimited piece.

Sequence

A sequence piece.

NamedSequence

A named sequence piece.

Methods

impl Specifier

fn specific_ident(ident: &str) -> Specifier

Returns a new Specifier corresponding to the given identifier.

fn specific_lftm(lftm: &str) -> Specifier

Returns a new Specifier corresponding to the given lifetime.

fn get_name(&self) -> Option<&String>

Returns the name of this specifier, if applicable.

fn to_fields(&self, context: &mut ExtCtxt, span: Span, stack: &[Amount]) -> Vec<Field>

Returns Fields that would initialize values matched by this specifier.

fn to_struct_fields(&self, context: &mut ExtCtxt, span: Span) -> Vec<StructField>

Returns StructFields that could contain values matched by this specifier.